Extracted all GM settings to a single OtherTab for now while preparing a GM-only tabbed settings dialog.
Moved info tab contents to its own directory in preparation for a standalone info dialog.
