+++ src/applications/settings/client-settings/tabs/NpcSettingsTab.svelte
@@ -10,48 +10,15 @@
     SettingsSheetContext,
     SettingsSheetFunctions,
   } from '../ClientSettings.types';
-  import ListboxSetting from '../parts/ListboxSetting.svelte';
-  import { CONSTANTS } from 'src/constants';
+  
 
   let context = getContext<Writable<SettingsSheetContext>>('context');
   let functions = getContext<SettingsSheetFunctions>('functions');
 
-  const userIsGm = FoundryAdapter.userIsGm();
-  const localize = FoundryAdapter.localize;
+    const localize = FoundryAdapter.localize;
 </script>
 
 <h2>{localize('TIDY5E.Settings.TabNPCs.header')}</h2>
-{#if userIsGm}
-  <SelectSetting
-    options={SettingsProvider.settings.initialNpcSheetTab.options.choices()}
-    bind:value={$context.settings.initialNpcSheetTab}
-    name={SettingsProvider.settings.initialNpcSheetTab.options.name}
-    hint={SettingsProvider.settings.initialNpcSheetTab.options.hint}
-    id="initialNpcSheetTab"
-  />
-
-  <ListboxSetting
-    name={SettingsProvider.settings.defaultNpcSheetTabs.options.name}
-    hint={SettingsProvider.settings.defaultNpcSheetTabs.options.hint}
-    leftHeader="TIDY5E.Settings.DefaultSheetTabs.AvailableHeader"
-    bind:leftItems={$context.defaultNpcTabs.available}
-    rightHeader="TIDY5E.Settings.DefaultSheetTabs.SelectedHeader"
-    bind:rightItems={$context.defaultNpcTabs.selected}
-    labelProp="label"
-    valueProp="id"
-  >
-    <div slot="below-listbox">
-      <button
-        type="button"
-        on:click={() =>
-          functions.resetDefaultTabs(context, CONSTANTS.SHEET_TYPE_NPC)}
-      >
-        <i class="fas fa-rotate-right" />
-        {localize('TIDY5E.Reset')}
-      </button>
-    </div>
-  </ListboxSetting>
-{/if}
 
 <CheckboxSetting
   bind:value={$context.settings.useClassicControlsForNpc}
@@ -59,34 +26,6 @@
   hint={SettingsProvider.settings.useClassicControlsForNpc.options.hint}
   id="useClassicControlsForNpc"
 />
-
-{#if userIsGm}
-  <CheckboxSetting
-    bind:value={$context.settings.useNpcRest}
-    name={'TIDY5E.Settings.UseNPCRest.name'}
-    hint={'TIDY5E.Settings.UseNPCRest.hint'}
-    id="useNpcRest"
-  />
-
-  <CheckboxSetting
-    bind:value={$context.settings.showNpcRestInChat}
-    name={'TIDY5E.Settings.ShowNPCRestInChat.name'}
-    hint={'TIDY5E.Settings.ShowNPCRestInChat.hint'}
-    id="showNpcRestInChat"
-  />
-
-  <SelectSetting
-    options={{
-      default: 'TIDY5E.Settings.ShowNPCActorLinkMarker.default',
-      unlinked: 'TIDY5E.Settings.ShowNPCActorLinkMarker.unlinked',
-      both: 'TIDY5E.Settings.ShowNPCActorLinkMarker.both',
-    }}
-    bind:value={$context.settings.showNpcActorLinkMarker}
-    name="TIDY5E.Settings.ShowNPCActorLinkMarker.name"
-    hint="TIDY5E.Settings.ShowNPCActorLinkMarker.hint"
-    id="showNpcActorLinkMarker"
-  />
-{/if}
 
 <CheckboxSetting
   bind:value={$context.settings.useHpBarNpc}
