+++ src/applications/settings/info/Info.svelte
