+++ src/applications/settings/world-settings/WorldSettings.types.ts
@@ -1,0 +1,11 @@
+import type { Readable } from 'svelte/store';
+
+export type WorldSettingsContext = {
+
+};
+
+export type WorldSettingsFunctions = {
+
+};
+
+export type WorldSettingsContextStore = Readable<WorldSettingsContext>;
